refactor: replace Langevin equation benchmark with Brownian motion MSD benchmarks
Replace Langevin equation simulation benchmark with focused MSD (mean squared displacement) benchmarks for Brownian motion. Add CPU (f64) and CUDA GPU (f32) MSD benchmarks using `Bm::default()`. Remove single trajectory simulation benchmark and Langevin equation setup. Update imports to use `continuous::Bm` instead of `langevin!` macro.
